requirements.txt
# Core Dependencies
google-generativeai>=0.3.0

# For API key management
python-dotenv>=0.19.0

# For async operations
asyncio>=3.4.3
aiofiles>=0.8.0

# For type hints (if not using Python 3.9+)
typing-extensions>=4.0.0

# For logging and error handling
loguru>=0.6.0
